fix(cloudfront): Set MinimumProtocolVersion and SslSupportMethod when specifying distribution certificate (#9200)
Per the CloudFormation documentation, if you specify an ACM certificate ARN, you must also specify values for
MinimumProtocolVersion and SslSupportMethod in AWS::CloudFront::Distribution ViewerCertificate. We are using
the recommended SslSupportMethod of "sni-only" and MinimumProtocolVersion of "TLSv1.2_2018."

Fixes: https://github.com/aws/aws-cdk/issues/9193

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*
