+++ packages/@aws-cdk/aws-cloudfront/test/distribution.test.ts
@@ -217,6 +217,8 @@ describe('certificates', () => {
       DistributionConfig: {
         ViewerCertificate: {
           AcmCertificateArn: 'arn:aws:acm:us-east-1:123456789012:certificate/12345678-1234-1234-1234-123456789012',
+          SslSupportMethod: 'sni-only',
+          MinimumProtocolVersion: 'TLSv1.2_2018',
         },
       },
     });
